Koin DI; Refactor project structure; Bump js version

+++ app/src/main/java/com/paulcoding/hviewer/ui/page/downloads/DownloadsPage.kt
@@ -5,6 +5,7 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
@@ -36,8 +37,8 @@ import com.paulcoding.hviewer.ui.component.ConfirmDialog
 import com.paulcoding.hviewer.ui.component.HBackIcon
 import com.paulcoding.hviewer.ui.component.HEmpty
 import com.paulcoding.hviewer.ui.component.HIcon
-import com.paulcoding.hviewer.ui.page.post.ImageModal
-import com.paulcoding.hviewer.ui.page.post.PostImage
+import com.paulcoding.hviewer.ui.page.sites.post.ImageModal
+import com.paulcoding.hviewer.ui.page.sites.post.PostImage
 import java.io.File
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -104,22 +105,24 @@ fun DownloadsPage(
                 ImageList(selectedDir!!)
             }
         }
-        ConfirmDialog(
-            showDialog = dirWillBeDeleted != null,
-            title = stringResource(R.string.confirm_delete),
-            text = stringResource(
-                R.string.are_you_sure_you_want_to_delete_folder,
-                dirWillBeDeleted?.name ?: ""
-            ),
-            onDismiss = {
-                dirWillBeDeleted = null
-            }, onConfirm = {
-                if (dirWillBeDeleted?.deleteRecursively() == true) {
-                    makeToast("Deleted ${dirWillBeDeleted?.name}")
-                    fetchDirs()
+        if (dirWillBeDeleted != null) {
+            ConfirmDialog(
+                title = stringResource(R.string.confirm_delete),
+                text = stringResource(
+                    R.string.are_you_sure_you_want_to_delete_folder,
+                    dirWillBeDeleted?.name ?: ""
+                ),
+                onDismiss = {
                     dirWillBeDeleted = null
+                }, onConfirm = {
+                    if (dirWillBeDeleted?.deleteRecursively() == true) {
+                        makeToast("Deleted ${dirWillBeDeleted?.name}")
+                        fetchDirs()
+                        dirWillBeDeleted = null
+                    }
                 }
-            })
+            )
+        }
     }
 }
 
@@ -153,8 +156,8 @@ internal fun ImageList(selectedDir: File) {
             HEmpty()
         }
     }
-    if (selectedImage != null) {
-        ImageModal(url = selectedImage!!) {
+    selectedImage?.let {
+        ImageModal(url = it) {
             selectedImage = null
         }
     }
